Refactor: modularize response boxes
